Modification so its a lib

+++ Lib/PTKeyBroadcaster.m
@@ -16,7 +16,7 @@ NSString* PTKeyBroadcasterKeyEvent = @"PTKeyBroadcasterKeyEvent";
 
 - (void)_bcastKeyCode: (short)keyCode modifiers: (long)modifiers
 {
-	PTKeyCombo* keyCombo = [PTKeyCombo keyComboWithKeyCode: keyCode modifiers: modifiers];
+	PTKeyCombo* keyCombo = [PTKeyCombo keyComboWithKeyCode: keyCode modifiers: (int)modifiers];
 	NSDictionary* userInfo = [NSDictionary dictionaryWithObject: keyCombo forKey:@"keyCombo"];
 
 	[[NSNotificationCenter defaultCenter]
